Make MonotonicClock not implement Clock and move to package private.
