+++ sdk/trace/src/test/java/io/opentelemetry/sdk/trace/RecordEventsReadableSpanTest.java
@@ -928,7 +928,7 @@ class RecordEventsReadableSpanTest {
             Context.root(),
             config,
             spanProcessor,
-            testClock,
+            MonotonicClock.create(testClock),
             resource,
             attributes,
             links,
@@ -1016,7 +1016,7 @@ class RecordEventsReadableSpanTest {
             Context.root(),
             spanLimits,
             spanProcessor,
-            clock,
+            MonotonicClock.create(clock),
             resource,
             attributesWithCapacity,
             Collections.singletonList(link1),
